Add files via upload
This is an important file for building a lobbying database that is searchable by the legislative title, e.g. "SB 50", "AB 32", etc. 

These names are buried as a string list in the LPAY file, and often this data needs hefty cleanup. I used regex to help clean the data to build a new SQL table of the legislation listed in the lobbying files.